+++ kommons-test/src/jvmMain/kotlin/com/bkahlert/kommons/test/junit/DynamicTestBuilder.kt
@@ -1,5 +1,6 @@
 package com.bkahlert.kommons.test.junit
 
+import com.bkahlert.kommons.logging.SLF4J
 import com.bkahlert.kommons.test.KommonsTest
 import com.bkahlert.kommons.test.junit.DynamicTestDisplayNameGenerator.FOR
 import com.bkahlert.kommons.test.junit.DynamicTestDisplayNameGenerator.assertingDisplayName
@@ -7,19 +8,30 @@ import com.bkahlert.kommons.test.junit.DynamicTestDisplayNameGenerator.catchingD
 import com.bkahlert.kommons.test.junit.DynamicTestDisplayNameGenerator.displayNameFor
 import com.bkahlert.kommons.test.junit.DynamicTestDisplayNameGenerator.expectingDisplayName
 import com.bkahlert.kommons.test.junit.DynamicTestDisplayNameGenerator.throwingDisplayName
+import com.bkahlert.kommons.test.junit.Mode.CREATE
+import com.bkahlert.kommons.test.junit.Mode.REPLACE
 import com.bkahlert.kommons.test.junit.PathSource.Companion.sourceUri
-import com.bkahlert.kommons.test.junit.SimpleIdResolver.Companion.simpleId
 import io.kotest.assertions.asClue
-import io.kotest.assertions.throwables.shouldThrow
+import io.kotest.assertions.assertionCounter
+import io.kotest.assertions.failure
 import io.kotest.assertions.withClue
+import io.kotest.mpp.bestName
 import org.junit.jupiter.api.DynamicContainer
 import org.junit.jupiter.api.DynamicContainer.dynamicContainer
 import org.junit.jupiter.api.DynamicNode
 import org.junit.jupiter.api.DynamicTest.dynamicTest
-import org.junit.jupiter.api.extension.AfterEachCallback
-import org.junit.jupiter.api.extension.ExtensionContext
 import java.net.URI
 import java.util.stream.Stream
+import kotlin.coroutines.Continuation
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.EmptyCoroutineContext
+import kotlin.coroutines.intrinsics.COROUTINE_SUSPENDED
+import kotlin.coroutines.intrinsics.createCoroutineUnintercepted
+import kotlin.coroutines.intrinsics.suspendCoroutineUninterceptedOrReturn
+import kotlin.coroutines.resume
+import kotlin.reflect.KClass
+import kotlin.streams.asStream
+
 
 /** Assertions that can be applied to a subject. */
 public typealias Assertions<T> = (T) -> Unit
@@ -28,25 +40,23 @@ public typealias Assertions<T> = (T) -> Unit
 @JvmInline
 public value class AssertionsBuilder<T>(
     /** Function that can be used to verify assertions passed to [it] or [that]. */
-    public val applyAssertions: (Assertions<T>) -> Unit,
+    public val applyAssertions: suspend (Assertions<T>) -> Unit,
 ) {
+
     /** Specifies [assertions] with the subject in the receiver `this`. */
-    public infix fun it(assertions: T.() -> Unit): Unit = applyAssertions(assertions)
+    public suspend infix fun it(assertions: T.() -> Unit): Unit = applyAssertions(assertions)
 
     /** Specifies [assertions] with the subject passed the single parameter `it`. */
-    public infix fun that(assertions: (T) -> Unit): Unit = applyAssertions(assertions)
+    public suspend infix fun that(assertions: (T) -> Unit): Unit = applyAssertions(assertions)
 }
 
 
-/** Builds tests with no subjects using a [DynamicTestsWithoutSubjectBuilder]. */
-public fun testing(init: DynamicTestsWithoutSubjectBuilder.() -> Unit): Stream<DynamicNode> =
-    DynamicTestsWithoutSubjectBuilder.build(init)
+/** Builds tests with no subjects using a [TestsWithoutSubjectScope]. */
+public fun testing(init: suspend TestsWithoutSubjectScope.() -> Unit): Stream<DynamicNode> =
+    buildTestNodeSequence(init) { DynamicTestsWithoutSubjectBuilder(it) }.asNonEmptyStream()
 
-/** Builder for tests (and test containers) with no subjects. */
-public class DynamicTestsWithoutSubjectBuilder(
-    public val addDynamicNode: (DynamicNode) -> Unit,
-) {
-
+/** Scope for building tests (and test containers) with no subjects. */
+public interface TestsWithoutSubjectScope {
     /**
      * Expects the subject returned by [action] to fulfil the
      * [Assertions] returned by [AssertionsBuilder].
@@ -55,20 +65,7 @@ public class DynamicTestsWithoutSubjectBuilder(
      *
      * **Usage:** `expecting { <action> } that { it.<assertions> }`
      */
-    public fun <R> expecting(description: String? = null, action: () -> R): AssertionsBuilder<R> {
-        var additionalAssertions: Assertions<R>? = null
-        val caller = KommonsTest.locateCall()
-        val test = dynamicTest(description ?: caller.expectingDisplayName(action), caller.sourceUri) {
-            additionalAssertions?.also {
-                val subject = action()
-                subject.asClue(it)
-            } ?: throw IllegalUsageException("expecting", caller.sourceUri)
-        }
-        addDynamicNode(test)
-        return AssertionsBuilder { assertions: Assertions<R> ->
-            additionalAssertions = assertions
-        }
-    }
+    public suspend fun <R> expecting(description: String? = null, action: () -> R): AssertionsBuilder<R>
 
     /**
      * Expects the [Result] returned by [action] to fulfil the
@@ -78,20 +75,7 @@ public class DynamicTestsWithoutSubjectBuilder(
      *
      * **Usage:** `expectCatching { <action> } that { it.<assertions> }`
      */
-    public fun <R> expectCatching(action: () -> R): AssertionsBuilder<Result<R>> {
-        var additionalAssertions: Assertions<Result<R>>? = null
-        val caller = KommonsTest.locateCall()
-        val test = dynamicTest(caller.catchingDisplayName(action), caller.sourceUri) {
-            additionalAssertions?.also {
-                val subject = runCatching(action)
-                subject.asClue(it)
-            } ?: throw IllegalUsageException("expectCatching", caller.sourceUri)
-        }
-        addDynamicNode(test)
-        return AssertionsBuilder { assertions: Assertions<Result<R>> ->
-            additionalAssertions = assertions
-        }
-    }
+    public suspend fun <R> expectCatching(action: () -> R): AssertionsBuilder<Result<R>>
 
     /**
      * Expects an exception [E] to be thrown when running [action]
@@ -103,35 +87,101 @@ public class DynamicTestsWithoutSubjectBuilder(
      *
      * **Usage:** `expectThrows<Exception> { <action> } that { it.<assertions> }`
      */
-    public inline fun <reified E : Throwable> expectThrows(noinline action: () -> Any?): AssertionsBuilder<E> {
-        var additionalAssertions: Assertions<E>? = null
+    public suspend fun <E : Throwable> expectThrows(type: KClass<E>, action: () -> Any?): AssertionsBuilder<E>
+}
+
+/**
+ * Expects an exception [E] to be thrown when running [action]
+ * and to optionally fulfil the [Assertions] returned by [AssertionsBuilder].
+ *
+ * **Usage:** `expectThrows<Exception> { <action> }`
+ *
+ * **Usage:** `expectThrows<Exception> { <action> } it { <assertions> }`
+ *
+ * **Usage:** `expectThrows<Exception> { <action> } that { it.<assertions> }`
+ */
+public suspend inline fun <reified E : Throwable> TestsWithoutSubjectScope.expectThrows(noinline action: () -> Any?): AssertionsBuilder<E> =
+    expectThrows(E::class, action)
+
+
+/** Builder for tests (and test containers) with no subjects. */
+private class DynamicTestsWithoutSubjectBuilder(
+    val yieldNode: suspend (Mode, DynamicNode) -> Unit,
+) : TestsWithoutSubjectScope {
+
+    override suspend fun <R> expecting(description: String?, action: () -> R): AssertionsBuilder<R> {
         val caller = KommonsTest.locateCall()
-        val test = dynamicTest(throwingDisplayName(E::class), caller.sourceUri) {
-            shouldThrow<E>(action).asClue(additionalAssertions ?: {})
-        }
-        addDynamicNode(test)
-        return AssertionsBuilder { assertions: Assertions<E> ->
-            additionalAssertions = assertions
+        val displayName = description ?: caller.expectingDisplayName(action)
+        yieldNode(CREATE, dynamicTest(displayName, caller.sourceUri) { throw IllegalUsageException("expecting", caller.sourceUri) })
+        return AssertionsBuilder { assertions: Assertions<R> ->
+            yieldNode(REPLACE, dynamicTest(displayName, caller.sourceUri) { action().asClue(assertions) })
         }
     }
 
-    public companion object {
-        public inline fun build(
-            init: DynamicTestsWithoutSubjectBuilder.() -> Unit,
-        ): Stream<DynamicNode> = buildList {
-            DynamicTestsWithoutSubjectBuilder { add(it) }.init()
-        }.stream()
+    override suspend fun <R> expectCatching(action: () -> R): AssertionsBuilder<Result<R>> {
+        val caller = KommonsTest.locateCall()
+        val displayName = caller.catchingDisplayName(action)
+        yieldNode(CREATE, dynamicTest(displayName, caller.sourceUri) { throw IllegalUsageException("expectCatching", caller.sourceUri) })
+        return AssertionsBuilder { assertions: Assertions<Result<R>> ->
+            yieldNode(REPLACE, dynamicTest(displayName, caller.sourceUri) { runCatching(action).asClue(assertions) })
+        }
+    }
+
+    override suspend fun <E : Throwable> expectThrows(type: KClass<E>, action: () -> Any?): AssertionsBuilder<E> {
+        val caller = KommonsTest.locateCall()
+        val displayName = throwingDisplayName(type)
+        yieldNode(CREATE, dynamicTest(displayName, caller.sourceUri) { shouldThrow(type, action).asClue({}) })
+        return AssertionsBuilder { assertions: Assertions<E> ->
+            yieldNode(REPLACE, dynamicTest(displayName, caller.sourceUri) { shouldThrow(type, action).asClue(assertions) })
+        }
     }
 }
 
 /**
- * Builds tests with the specified [subject] using a [DynamicTestsWithSubjectBuilder].
+ * Copy of [io.kotest.assertions.throwables.shouldThrow] that allows passing
+ * the [expectedExceptionClass] as a [KClass] instance.
  */
-public fun <T> testing(subject: T, init: DynamicTestsWithSubjectBuilder<T>.() -> Unit): Stream<DynamicNode> =
-    DynamicTestsWithSubjectBuilder.build(subject, init)
+private fun <T : Throwable> shouldThrow(expectedExceptionClass: KClass<T>, block: () -> Any?): T {
+    assertionCounter.inc()
+    val thrownThrowable = try {
+        block()
+        null  // Can't throw failure here directly, as it would be caught by the catch clause, and it's an AssertionError, which is a special case
+    } catch (thrown: Throwable) {
+        thrown
+    }
+
+    @Suppress("UNCHECKED_CAST")
+    return when {
+        thrownThrowable == null -> {
+            throw failure("Expected exception ${expectedExceptionClass.bestName()} but no exception was thrown.")
+        }
+
+        expectedExceptionClass.isInstance(thrownThrowable) -> {
+            // This should be before `is AssertionError`. If the user is purposefully trying to verify `shouldThrow<AssertionError>{}` this will take priority
+            thrownThrowable as T
+        }
+
+        thrownThrowable is AssertionError -> {
+            throw thrownThrowable
+        }
+
+        else -> {
+            throw failure(
+                "Expected exception ${expectedExceptionClass.bestName()} but a ${thrownThrowable::class.simpleName} was thrown instead.",
+                thrownThrowable
+            )
+        }
+    }
+}
 
 /**
- * Builds tests for each of the specified [subjects] using a [DynamicTestsWithSubjectBuilder].
+ * Builds tests with the specified [subject] using a [TestsWithSubjectScope].
+ */
+public fun <T> testing(subject: T, init: suspend TestsWithSubjectScope<T>.() -> Unit): Stream<DynamicNode> =
+    buildTestNodeSequence(init) { DynamicTestsWithSubjectBuilder(subject, it) }.asNonEmptyStream()
+
+/**
+ * Builds tests for each of the specified [subjects] using a [TestsWithSubjectScope].
  *
  * The name for each container is heuristically derived but can also be explicitly specified using [containerNamePattern]
  * which supports curly placeholders `{}` like [SLF4J] does.
@@ -139,134 +189,86 @@ public fun <T> testing(subject: T, init: DynamicTestsWithSubjectBuilder<T>.() ->
 public fun <T> testingAll(
     vararg subjects: T,
     containerNamePattern: String? = null,
-    init: DynamicTestsWithSubjectBuilder<T>.() -> Unit,
-): Stream<DynamicContainer> = subjects.asList().testingAll(containerNamePattern, init)
+    init: suspend TestsWithSubjectScope<T>.() -> Unit,
+): Stream<DynamicContainer> = subjects.asIterable().testingAll(containerNamePattern, init)
 
 /**
- * Builds tests for each of subject of this [Collection] using a [DynamicTestsWithSubjectBuilder].
+ * Builds tests for each subject of this [Collection] using a [TestsWithSubjectScope].
  *
  * The name for each container is heuristically derived but can also be explicitly specified using [containerNamePattern]
  * which supports curly placeholders `{}` like [SLF4J] does.
  */
 public fun <T> Iterable<T>.testingAll(
     containerNamePattern: String? = null,
-    init: DynamicTestsWithSubjectBuilder<T>.() -> Unit,
-): Stream<DynamicContainer> = toList()
-    .also { require(it.isNotEmpty()) { "At least one subject must be provided for testing." } }
-    .map { subject ->
-        dynamicContainer(
-            "$FOR ${displayNameFor(subject, containerNamePattern)}",
-            PathSource.currentUri,
-            testing(subject, init)
-        )
-    }.stream()
+    init: suspend TestsWithSubjectScope<T>.() -> Unit,
+): Stream<DynamicContainer> = map { subject ->
+    dynamicContainer(
+        "$FOR ${displayNameFor(subject, containerNamePattern)}",
+        PathSource.currentUri,
+        testing(subject, init)
+    )
+}.asSequence().asNonEmptyStream()
 
 /**
- * Builds tests for each of subject of this [Sequence] using a [DynamicTestsWithSubjectBuilder].
+ * Builds tests for each subject of this [Sequence] using a [TestsWithSubjectScope].
  *
  * The name for each container is heuristically derived but can also be explicitly specified using [containerNamePattern]
  * which supports curly placeholders `{}` like [SLF4J] does.
  */
 public fun <T> Sequence<T>.testingAll(
     containerNamePattern: String? = null,
-    init: DynamicTestsWithSubjectBuilder<T>.() -> Unit,
-): Stream<DynamicContainer> = toList().testingAll(containerNamePattern, init)
+    init: TestsWithSubjectScope<T>.() -> Unit,
+): Stream<DynamicContainer> = asIterable().testingAll(containerNamePattern, init)
 
 /**
- * Builds tests for each of entry of this [Map] using a [DynamicTestsWithSubjectBuilder].
+ * Builds tests for each entry of this [Map] using a [TestsWithSubjectScope].
  *
  * The name for each container is heuristically derived but can also be explicitly specified using [containerNamePattern]
  * which supports curly placeholders `{}` like [SLF4J] does.
  */
 public fun <K, V> Map<K, V>.testingAll(
     containerNamePattern: String? = null,
-    init: DynamicTestsWithSubjectBuilder<Map.Entry<K, V>>.() -> Unit,
+    init: TestsWithSubjectScope<Map.Entry<K, V>>.() -> Unit,
 ): Stream<DynamicContainer> = entries.testingAll(containerNamePattern, init)
 
-/** Builder for tests (and test containers) with the specified [subject]. */
-public class DynamicTestsWithSubjectBuilder<T>(
-    public val subject: T,
-    public val addDynamicNode: (DynamicNode) -> Unit,
-) {
-
+/** Scope for building tests (and test containers) with a subject. */
+public interface TestsWithSubjectScope<T> {
     /**
-     * Expects the [subject] to fulfil the given [assertions].
+     * Expects the subject to fulfil the given [assertions].
      *
      * **Usage:** `it { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<assertions> }`
      */
-    public fun it(assertions: T.() -> Unit) {
-        val caller = KommonsTest.locateCall()
-        val test = dynamicTest(caller.assertingDisplayName(subject, assertions), caller.sourceUri) {
-            subject.asClue(assertions)
-        }
-        addDynamicNode(test)
-    }
+    public suspend fun it(assertions: T.() -> Unit)
 
     /**
-     * Expects the [subject] to fulfil the given [assertions].
+     * Expects the subject to fulfil the given [assertions].
      *
      * **Usage:** `that { it.<assertions> }`
      */
-    public fun that(assertions: Assertions<T>) {
-        val caller = KommonsTest.locateCall()
-        val test = dynamicTest(caller.assertingDisplayName(subject, assertions), caller.sourceUri) {
-            subject.asClue(assertions)
-        }
-        addDynamicNode(test)
-    }
+    public suspend fun that(assertions: Assertions<T>)
 
     /**
-     * Expects the [subject] transformed by [action] to fulfil the
+     * Expects the subject transformed by [action] to fulfil the
      * [Assertions] returned by [AssertionsBuilder].
      *
      * **Usage:** `expecting { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<action> } it { <assertions> }`
      *
      * **Usage:** `expecting { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<action> } that { it.<assertions> }`
      */
-    public fun <R> expecting(description: String? = null, action: T.() -> R): AssertionsBuilder<R> {
-        var additionalAssertions: Assertions<R>? = null
-        val caller = KommonsTest.locateCall()
-        val test = dynamicTest(description ?: caller.expectingDisplayName(action), caller.sourceUri) {
-            additionalAssertions?.also {
-                withClue(subject) {
-                    val aspect = subject.action()
-                    aspect.asClue(it)
-                }
-            } ?: throw IllegalUsageException("expecting", caller.sourceUri)
-        }
-        addDynamicNode(test)
-        return AssertionsBuilder { assertions: Assertions<R> ->
-            additionalAssertions = assertions
-        }
-    }
+    public suspend fun <R> expecting(description: String? = null, action: T.() -> R): AssertionsBuilder<R>
 
     /**
-     * Expects the [Result] of the [subject] transformed by [action] to fulfil the
+     * Expects the [Result] of the subject transformed by [action] to fulfil the
      * [Assertions] returned by [AssertionsBuilder].
      *
      * **Usage:** `expectCatching { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<action> } it { <assertions> }`
      *
      * **Usage:** `expectCatching { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<action> } that { it.<assertions> }`
      */
-    public fun <R> expectCatching(action: T.() -> R): AssertionsBuilder<Result<R>> {
-        var additionalAssertions: Assertions<Result<R>>? = null
-        val caller = KommonsTest.locateCall()
-        val test = dynamicTest(caller.catchingDisplayName(action), caller.sourceUri) {
-            additionalAssertions?.also {
-                withClue(subject) {
-                    val aspect = subject.runCatching(action)
-                    aspect.asClue(it)
-                }
-            } ?: throw IllegalUsageException("expectCatching", caller.sourceUri)
-        }
-        addDynamicNode(test)
-        return AssertionsBuilder { assertions: Assertions<Result<R>> ->
-            additionalAssertions = assertions
-        }
-    }
+    public suspend fun <R> expectCatching(action: T.() -> R): AssertionsBuilder<Result<R>>
 
     /**
-     * Expects an exception [E] to be thrown when transforming the [subject] with [action]
+     * Expects an exception [E] to be thrown when transforming the subject with [action]
      * and to optionally fulfil the [Assertions] returned by [AssertionsBuilder].
      *
      * **Usage:** `expectThrows<Exception> { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<action> }`
@@ -275,60 +277,250 @@ public class DynamicTestsWithSubjectBuilder<T>(
      *
      * **Usage:** `expectThrows<Exception> { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<action> } that { it.<assertions> }`
      */
-    public inline fun <reified E : Throwable> expectThrows(noinline action: T.() -> Any?): AssertionsBuilder<E> {
-        var additionalAssertions: Assertions<E>? = null
+    public suspend fun <E : Throwable> expectThrows(type: KClass<E>, action: T.() -> Any?): AssertionsBuilder<E>
+}
+
+/**
+ * Expects an exception [E] to be thrown when transforming the subject with [action]
+ * and to optionally fulfil the [Assertions] returned by [AssertionsBuilder].
+ *
+ * **Usage:** `expectThrows<Exception> { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<action> }`
+ *
+ * **Usage:** `expectThrows<Exception> { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<action> } it { <assertions> }`
+ *
+ * **Usage:** `expectThrows<Exception> { 𝘴𝘶𝘣𝘫𝘦𝘤𝘵.<action> } that { it.<assertions> }`
+ */
+public suspend inline fun <reified E : Throwable, T> TestsWithSubjectScope<T>.expectThrows(noinline action: T.() -> Any?): AssertionsBuilder<E> =
+    expectThrows(E::class, action)
+
+
+/** Builder for tests (and test containers) with the specified [subject]. */
+private class DynamicTestsWithSubjectBuilder<T>(
+    val subject: T,
+    val yieldNode: suspend (Mode, DynamicNode) -> Unit,
+) : TestsWithSubjectScope<T> {
+
+    override suspend fun it(assertions: T.() -> Unit) {
         val caller = KommonsTest.locateCall()
-        val test = dynamicTest(throwingDisplayName(E::class), caller.sourceUri) {
-            withClue(subject) {
-                shouldThrow<E> { subject.action() }.asClue(additionalAssertions ?: {})
-            }
-        }
-        addDynamicNode(test)
-        return AssertionsBuilder { assertions: Assertions<E> ->
-            additionalAssertions = assertions
+        val displayName = caller.assertingDisplayName(subject, assertions)
+        yieldNode(CREATE, dynamicTest(displayName, caller.sourceUri) { subject.asClue(assertions) })
+    }
+
+    override suspend fun that(assertions: Assertions<T>) {
+        val caller = KommonsTest.locateCall()
+        val displayName = caller.assertingDisplayName(subject, assertions)
+        yieldNode(CREATE, dynamicTest(displayName, caller.sourceUri) { subject.asClue(assertions) })
+    }
+
+    override suspend fun <R> expecting(description: String?, action: T.() -> R): AssertionsBuilder<R> {
+        val caller = KommonsTest.locateCall()
+        val displayName = description ?: caller.expectingDisplayName(action)
+        yieldNode(CREATE, dynamicTest(displayName, caller.sourceUri) { throw IllegalUsageException("expecting", caller.sourceUri) })
+        return AssertionsBuilder { assertions: Assertions<R> ->
+            yieldNode(REPLACE, dynamicTest(displayName, caller.sourceUri) {
+                withClue(subject) { subject.action().asClue(assertions) }
+            })
         }
     }
 
-    public companion object {
+    override suspend fun <R> expectCatching(action: T.() -> R): AssertionsBuilder<Result<R>> {
+        val caller = KommonsTest.locateCall()
+        val displayName = caller.catchingDisplayName(action)
+        yieldNode(CREATE, dynamicTest(displayName, caller.sourceUri) { throw IllegalUsageException("expectCatching", caller.sourceUri) })
+        return AssertionsBuilder { assertions: Assertions<Result<R>> ->
+            yieldNode(REPLACE, dynamicTest(displayName, caller.sourceUri) {
+                withClue(subject) { subject.runCatching(action).asClue(assertions) }
+            })
+        }
+    }
 
-        public fun <T> build(
-            subject: T,
-            init: DynamicTestsWithSubjectBuilder<T>.() -> Unit,
-        ): Stream<DynamicNode> = buildList {
-            DynamicTestsWithSubjectBuilder(subject) { add(it) }.init()
-        }.stream()
+    override suspend fun <E : Throwable> expectThrows(type: KClass<E>, action: T.() -> Any?): AssertionsBuilder<E> {
+        val caller = KommonsTest.locateCall()
+        val displayName = throwingDisplayName(type)
+        yieldNode(CREATE, dynamicTest(displayName, caller.sourceUri) {
+            withClue(subject) { shouldThrow(type) { subject.action() }.asClue({}) }
+        })
+        return AssertionsBuilder { assertions: Assertions<E> ->
+            yieldNode(REPLACE, dynamicTest(displayName, caller.sourceUri) {
+                withClue(subject) { shouldThrow(type) { subject.action() }.asClue(assertions) }
+            })
+        }
     }
 }
 
 
-/** Exception thrown if the API is incorrectly used. */
+/** Exception that is thrown if the API is incorrectly used. */
 public class IllegalUsageException(function: String, caller: URI?) : IllegalArgumentException(
     "$function { … } call was not finished with \"that { … }\"".let {
         caller?.let { uri -> "$it at " + uri.path + ":" + uri.query.takeLastWhile { it.isDigit() } } ?: it
     }
 )
 
+private fun <T : DynamicNode> Sequence<T>.asNonEmptyStream() =
+    ifEmpty { throw IllegalStateException("No tests were created.") }.asStream()
+
+
+private enum class Mode { CREATE, REPLACE }
 
 /**
- * Extension that checks if [DynamicTestsWithSubjectBuilder.expecting] or [DynamicTestsWithSubjectBuilder.expectCatching]
- * where incorrectly used.
+ * Builds a sequence of dynamic nodes using [init]
+ * operating on the scope/builder [S].
  *
- * ***Important:**
- * For this extension to work, it needs to be registered.*
- *
- * > The most convenient way to register this extension
- * > for all tests is by adding the line **`com.bkahlert.kommons.test.junit.IllegalUsageCheck`** to the
- * > file **`resources/META-INF/services/org.junit.jupiter.api.extension.Extension`**.
+ * [initBuilder] has to provide a new builder
+ * that can use the passed suspend function to yield
+ * dynamic nodes and if necessary, [REPLACE] the
+ * previously yielded one.
  */
-internal class IllegalUsageCheck : AfterEachCallback {
+private fun <S> buildTestNodeSequence(
+    init: suspend S.() -> Unit,
+    initBuilder: (suspend (Mode, DynamicNode) -> Unit) -> S,
+): Sequence<DynamicNode> = unrestrictedSequence<DynamicNode> {
+    var scheduledNode: DynamicNode? = null
+    initBuilder { mode, newNode ->
+        when (mode) {
+            CREATE -> {
+                scheduledNode?.also { yield(it) }
+                scheduledNode = newNode
+            }
 
-    override fun afterEach(context: ExtensionContext) {
-        val id: SimpleId = context.simpleId
-        val illegalUsage = illegalUsages[id]
-        if (illegalUsage != null) throw illegalUsage
+            REPLACE -> {
+                scheduledNode?.also {
+                    if (!it.testSourceUri.equals(newNode.testSourceUri)) {
+                        throw IllegalStateException("${newNode.testSourceUri} attempts to replace the supposedly not fully built test ${it.testSourceUri}")
+                    }
+                }
+                scheduledNode = newNode
+            }
+        }
+    }.init()
+    scheduledNode?.also { yield(it) }
+}
+
+
+/**
+ * Simplified copy of [kotlin.sequences.SequenceScope] to
+ * support building tests while streaming them.
+ */
+private abstract class SequenceScope<in T> {
+    /** @see kotlin.sequences.SequenceScope.yield */
+    abstract suspend fun yield(value: T)
+
+    /** @see kotlin.sequences.SequenceScope.yieldAll */
+    abstract suspend fun yieldAll(iterator: Iterator<T>)
+
+    /** @see kotlin.sequences.SequenceScope.yieldAll */
+    suspend fun yieldAll(elements: Iterable<T>) {
+        if (elements is Collection && elements.isEmpty()) return
+        return yieldAll(elements.iterator())
     }
 
-    companion object {
-        val illegalUsages: MutableMap<SimpleId, IllegalUsageException> = mutableMapOf()
+    /** @see kotlin.sequences.SequenceScope.yieldAll */
+    suspend fun yieldAll(sequence: Sequence<T>): Unit = yieldAll(sequence.iterator())
+}
+
+private fun <T> unrestrictedSequence(@BuilderInference block: suspend SequenceScope<T>.() -> Unit): Sequence<T> = Sequence { unrestrictedIterator(block) }
+
+private fun <T> unrestrictedIterator(@BuilderInference block: suspend SequenceScope<T>.() -> Unit): Iterator<T> {
+    val iterator = SequenceBuilderIterator<T>()
+    iterator.nextStep = block.createCoroutineUnintercepted(receiver = iterator, completion = iterator)
+    return iterator
+}
+
+private typealias State = Int
+
+private const val State_NotReady: State = 0
+private const val State_ManyNotReady: State = 1
+private const val State_ManyReady: State = 2
+private const val State_Ready: State = 3
+private const val State_Done: State = 4
+private const val State_Failed: State = 5
+
+private class SequenceBuilderIterator<T> : SequenceScope<T>(), Iterator<T>, Continuation<Unit> {
+    private var state = State_NotReady
+    private var nextValue: T? = null
+    private var nextIterator: Iterator<T>? = null
+    var nextStep: Continuation<Unit>? = null
+
+    override fun hasNext(): Boolean {
+        while (true) {
+            when (state) {
+                State_NotReady -> {}
+                State_ManyNotReady ->
+                    if (nextIterator!!.hasNext()) {
+                        state = State_ManyReady
+                        return true
+                    } else {
+                        nextIterator = null
+                    }
+
+                State_Done -> return false
+                State_Ready, State_ManyReady -> return true
+                else -> throw exceptionalState()
+            }
+
+            state = State_Failed
+            val step = nextStep!!
+            nextStep = null
+            step.resume(Unit)
+        }
     }
+
+    override fun next(): T {
+        when (state) {
+            State_NotReady, State_ManyNotReady -> return nextNotReady()
+            State_ManyReady -> {
+                state = State_ManyNotReady
+                return nextIterator!!.next()
+            }
+
+            State_Ready -> {
+                state = State_NotReady
+                @Suppress("UNCHECKED_CAST")
+                val result = nextValue as T
+                nextValue = null
+                return result
+            }
+
+            else -> throw exceptionalState()
+        }
+    }
+
+    private fun nextNotReady(): T {
+        if (!hasNext()) throw NoSuchElementException() else return next()
+    }
+
+    private fun exceptionalState(): Throwable = when (state) {
+        State_Done -> NoSuchElementException()
+        State_Failed -> IllegalStateException("Iterator has failed.")
+        else -> IllegalStateException("Unexpected state of the iterator: $state")
+    }
+
+
+    override suspend fun yield(value: T) {
+        nextValue = value
+        state = State_Ready
+        return suspendCoroutineUninterceptedOrReturn { c ->
+            nextStep = c
+            COROUTINE_SUSPENDED
+        }
+    }
+
+    override suspend fun yieldAll(iterator: Iterator<T>) {
+        if (!iterator.hasNext()) return
+        nextIterator = iterator
+        state = State_ManyReady
+        return suspendCoroutineUninterceptedOrReturn { c ->
+            nextStep = c
+            COROUTINE_SUSPENDED
+        }
+    }
+
+    // Completion continuation implementation
+    override fun resumeWith(result: Result<Unit>) {
+        result.getOrThrow() // just rethrow exception if it's there
+        state = State_Done
+    }
+
+    override val context: CoroutineContext
+        get() = EmptyCoroutineContext
 }